components/transport_mqtt/e2e_tests/requirements.txt
paho-mqtt==1.6.1
pytest==7.4.3
pytest-asyncio==0.21.1
pyserial==3.5
psutil==5.9.6
docker==6.1.3
requests==2.31.0